feat(ci): unccoment lint test and api unit test jobs
This commit adds lint-test-job and api-unit-tests-job to the .gitlab-ci.yml file. The lint-test-job installs flake8 and runs it on the 'api' directory. The api-unit-tests-job sets up a Python environment, installs requirements, and runs pytest on the 'api' directory.

build(api): build and push API Docker image

This commit modifies the .gitlab-ci.yml file to include the build-api-image job. It sets up a Docker image with git, logs into the GitLab container registry, builds the API image from the 'api' directory, and pushes it to the registry.

modify(api): update Dockerfile to include additional files

This commit updates the Dockerfile in the 'api' directory. It adds the 'migrations', 'queries', 'routers', 'authenticator.py', 'main.py', and 'requirements.txt' files to the Docker image.

fix(api): add root path endpoint in main.py

This commit adds a root path endpoint ("/") to the main.py file in the 'api' directory. The endpoint returns a JSON response with a message when accessed.